requirements.txt
Flask==2.3.2
tensorflow==2.13.0
Pillow==10.0.0
numpy==1.24.3